04-analytics-engineering/astro-airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-google
astronomer-cosmos[dbt-bigquery]
astronomer-cosmos[dbt-postgres]
pyarrow